# api/swagger.yaml
swagger: '2.0'
info:
  description: API for managing Travel Conceriage Service datasets.
  version: 1.0.0
  title: Travel Conseriage
  termsOfService: TBD
  contact:
    email: ammar.haris@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8080'
basePath: /tc
...
schemes:
  - http
paths:
...
  /agency:
    post:
      tags:
        - agency
      summary: Create Agency
      description: This can only be done by the logged in user.
      operationId: createAgency
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Created Agency object
          required: true
          schema:
            $ref: '#/definitions/Agency'
      responses:
        default:
          description: successful operation
  /agency/getAgencyById:
    get:
      tags:
        - agency
      summary: Returns an Agency for a given id
      description: Returns agency object
      operationId: getAgencyById
      produces:
        - application/json
      parameters:
        - name: agencyId
          in: query
          description: The id of the agency
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Agency'
      security:
        - api_key: []
  /agency/getAgency:
    get:
      tags:
        - agency
      summary: Returns all agencies in the system
      description: Returns list of all Agencies
      operationId: getAgencies
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Agency'
      security:
        - api_key: []
securityDefinitions:
  agent_auth:
    type: oauth2
    authorizationUrl: 'http://abc.com/oauth/dialog'
    flow: implicit
    scopes:
      'write:packages': modify packages in your account
      'read:packages': read your packages
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
...
  Agency:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      address1:
        type: string
      address2:
        type: string
      city:
        type: string
      state:
        type: string
      zip:
        type: string
      country:
        type: string
      Email:
        type: string
      Phone:
        type: string
      Website:
        type: string
      isDeleted:
        type: boolean
        default: false
    xml:
      name: Agent
...